requirements.txt
wheel
--find-links https://download.pytorch.org/whl/cu102/torch_stable.html
torch==1.7.0
torchvision==0.8.1
jupyterlab==2.2.9
matplotlib==2.2.4
Pillow==6.2.0
scikit-learn==0.22
scipy==1.2.1
numpy==1.16.0
protobuf==3.10.0
pandas==1.0.3
absl-py
scikit-image
six